程序能跑，修正许多bug 还剩部分错误题目已标志出 32%指令集中指令未涉及

+++ code/mips_sim/mips_sim/op.hpp
@@ -181,7 +181,7 @@ public:
 				temp = temp + (int)(Idata[1].location);
 			}
 			else {
-				temp = memRef[Idata[0].location];
+				temp = memRef[(unsigned int)Idata[0].location];
 			}
 			if (ins[0] == 'l') {
 				Odata.fromNumber((unsigned long long)temp);
